Create CreateCMPackages_BIOS_Drivers.ps1
Creates Packages based on the hash table in the script.  Change you models accordingly.
This creates the folders structure based on the model information, creates a couple text files, starts the distribution
This is the foundation script for the next script that then consumes the information in the CM Packages, and downloads directly from the vendors to populate the folder structure with current source and update CM Package fields.